fix(text): rename files in order to make more accurate to the context
