+++ src/components/data/Text/stories/Text.Paragraphs.stories.tsx
@@ -11,7 +11,7 @@ import {
   appearance,
   typo,
 } from "./props";
-import { ITextProps } from "../interfaces/TextProps.interface";
+import { ITextProps } from "../interfaces/Text.interface";
 
 const story = {
   title: "data/Text/Paragraphs",
