+++ src/components/data/Text/types/Text.AlignOptions.type.ts
