+++ src/components/data/Text/types/Text.TyposOptions.type.ts
